Reduce the store code
