+++ package.yaml
@@ -141,7 +141,6 @@ library:
   - Data.Attoparsec.Combinators
   - Data.Attoparsec.Interpreter
   - Data.Monoid.Map
-  - Data.Store.VersionTagged
   - Network.HTTP.Download
   - Network.HTTP.Download.Verified
   - Network.HTTP.StackClient
